serializer: Rename serialization dir to serializer
It's quite vivid for the nameing of `serializer`. I would like to
rename the top dir into `serializer` instead of `serialization` for
consistence with other module.
